backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.1.0
